claude-code/skills/video-to-summary/requirements.txt
# System dependencies (install with pip install --break-system-packages or use venv)
yt-dlp>=2024.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
pyyaml>=6.0.0
tqdm>=4.66.0

# Note: openai-whisper should be installed via pipx for better dependency management
# pipx install openai-whisper
